+++ src/coreclr/tools/aot/ILCompiler.ReadyToRun/ObjectWriter/ProfileFileBuilder.cs
@@ -157,7 +157,7 @@ namespace ILCompiler.PEWriter
                 if (_symbolMethodMap.TryGetValue(kvpCallerCalleeCount.Key, out ISymbolDefinitionNode callerSymbol) &&
                     _outputInfoBuilder.NodeSymbolMap.TryGetValue(callerSymbol, out callerNode))
                 {
-                    callerRVA = _outputInfoBuilder.Sections[callerNode.SectionIndex].VirtualAddress + (ulong)callerNode.Offset;
+                    callerRVA = _outputInfoBuilder.Sections[callerNode.SectionIndex].VirtualAddress + callerNode.Offset;
                 }
 
                 foreach (KeyValuePair<MethodDesc, int> kvpCalleeCount in kvpCallerCalleeCount.Value)
@@ -167,7 +167,7 @@ namespace ILCompiler.PEWriter
                     if (_symbolMethodMap.TryGetValue(kvpCalleeCount.Key, out ISymbolDefinitionNode calleeSymbol) &&
                         _outputInfoBuilder.NodeSymbolMap.TryGetValue(calleeSymbol, out calleeNode))
                     {
-                        calleeRVA = _outputInfoBuilder.Sections[calleeNode.SectionIndex].VirtualAddress + (ulong)calleeNode.Offset;
+                        calleeRVA = _outputInfoBuilder.Sections[calleeNode.SectionIndex].VirtualAddress + calleeNode.Offset;
                     }
 
                     _callInfo.Add(new CallInfo(
